Split job status checking by job type.
