+++ src/tools/async-download/types/asyncDownloadFormParameters.ts
@@ -1,7 +1,6 @@
 import { DownloadUrlOptions } from '../../../shared/config/apiUrls';
 
-export type FormParameters = DownloadUrlOptions & {
+export type FormParameters = Exclude<DownloadUrlOptions, 'accessions'> & {
   download: false;
   compressed: false;
-  accessions: undefined; // TODO: confirm
 };
