+++ src/tools/state/getCheckIdMappingJobStatus.ts
@@ -1,0 +1,138 @@
+import { Dispatch, MutableRefObject } from 'react';
+
+import fetchData from '../../shared/utils/fetchData';
+import { getStatusFromResponse, getJobMessage } from '../utils';
+import * as logging from '../../shared/utils/logging';
+
+import toolsURLs from '../config/urls';
+
+import { updateJob } from './toolsActions';
+import { addMessage } from '../../messages/state/messagesActions';
+
+import { ToolsState } from './toolsInitialState';
+import { ToolsAction } from './toolsReducers';
+import { MessagesAction } from '../../messages/state/messagesReducers';
+import { RunningJob, FinishedJob } from '../types/toolsJob';
+import { Status } from '../types/toolsStatuses';
+import { JobTypes } from '../types/toolsJobTypes';
+import { MappingError } from '../id-mapping/types/idMappingSearchResults';
+
+const possibleStatuses = new Set(Object.values(Status));
+
+const getCheckIdMappingJobStatus =
+  (
+    dispatch: Dispatch<ToolsAction>,
+    stateRef: MutableRefObject<ToolsState>,
+    messagesDispatch: Dispatch<MessagesAction>
+  ) =>
+  async (job: RunningJob | FinishedJob<JobTypes>) => {
+    const urlConfig = toolsURLs(job.type);
+    try {
+      // we use plain fetch as through Axios we cannot block redirects
+      const response = await window.fetch(urlConfig.statusUrl(job.remoteID), {
+        headers: {
+          Accept: 'text/plain,application/json',
+        },
+        method: 'GET',
+        redirect: 'follow',
+      });
+
+      const [status, idMappingResultsUrl] = await getStatusFromResponse(
+        job.type,
+        response
+      );
+
+      if (
+        !response.ok &&
+        status !== Status.FAILURE &&
+        status !== Status.ERRORED
+      ) {
+        throw new Error(`${response.status}: ${response.statusText}`);
+      }
+      // stateRef not hydrated yet
+      if (!stateRef.current) {
+        return;
+      }
+      // get a new reference to the job
+      let currentStateOfJob = stateRef.current[job.internalID];
+      // check that the job is still in the state (it might have been removed)
+      if (!currentStateOfJob) {
+        return;
+      }
+      // check that the status we got from the server is something expected
+      if (!possibleStatuses.has(status)) {
+        throw new Error(
+          `got an unexpected status of "${status}" from the server`
+        );
+      }
+      if (
+        status === Status.FINISHED &&
+        currentStateOfJob.status === Status.FINISHED
+      ) {
+        // job was already finished, and is still in the same state on the server
+        return;
+      }
+      if (status === Status.FAILURE) {
+        const errorResponse: { jobStatus: Status; errors?: MappingError[] } =
+          await response.json();
+        if (errorResponse.errors?.[0]) {
+          const error = errorResponse.errors[0];
+          dispatch(
+            updateJob(job.internalID, {
+              timeLastUpdate: Date.now(),
+              status,
+              errorDescription: error.message,
+            })
+          );
+        }
+      }
+      if (
+        status === Status.NOT_FOUND ||
+        status === Status.RUNNING ||
+        status === Status.FAILURE ||
+        status === Status.ERRORED
+      ) {
+        dispatch(
+          updateJob(job.internalID, {
+            timeLastUpdate: Date.now(),
+            status,
+          })
+        );
+        return;
+      }
+      // job finished, handle differently depending on job type
+
+      if (idMappingResultsUrl) {
+        // only ID Mapping jobs
+        const response = await fetchData(idMappingResultsUrl, undefined, {
+          method: 'HEAD',
+        });
+        // get a new reference to the job
+        currentStateOfJob = stateRef.current[job.internalID];
+        // check that the job is still in the state (it might have been removed)
+        if (!currentStateOfJob) {
+          return;
+        }
+        const now = Date.now();
+        const hits: string = response.headers['x-total-results'] || '0';
+        dispatch(
+          updateJob(job.internalID, {
+            timeLastUpdate: now,
+            timeFinished: now,
+            seen: false,
+            status,
+            data: { hits: +hits },
+          })
+        );
+        messagesDispatch(
+          addMessage(getJobMessage({ job: currentStateOfJob, nHits: +hits }))
+        );
+      }
+    } catch (error) {
+      if (error instanceof Error || typeof error === 'string') {
+        logging.error(error);
+      }
+    }
+  };
+
+export default getCheckIdMappingJobStatus;
